requirements.txt
Django==4.2.20
djangorestframework==3.15.2
psycopg2==2.9.10
djangorestframework-simplejwt==5.5.0
# celery[redis]==5.2.7
# flower==1.2.0
# celery_singleton==0.3.1
# django-redis==5.2.0
# django-cachalot==2.5.2
#